Authorization with Gate

+++ app/Http/Controllers/ArtistController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Artist;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class ArtistController extends Controller
 {
@@ -26,6 +27,10 @@ class ArtistController extends Controller
      */
     public function create()
     {
+        if (!Gate::allows('create-artist')) {
+            abort(403);
+        }
+
         return view('artist.create');
     }
 
@@ -72,6 +77,10 @@ class ArtistController extends Controller
      */
     public function edit(string $id)
     {
+        if (!Gate::allows('create-artist')) {
+            abort(403);
+        }
+
         $artist = Artist::find($id);
 
         return view('artist.edit', [
@@ -108,6 +117,10 @@ class ArtistController extends Controller
      */
     public function destroy(string $id)
     {
+        if (!Gate::allows('create-artist')) {
+            abort(403);
+        }
+        
         Artist::destroy($id);
 
         return redirect()->route('artist.index');
